Use `Arc<FileDecryptionProperties>` to reduce size of ParquetMetadata and avoid copying when `encryption` is enabled (#8470)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes https://github.com/apache/arrow-rs/issues/8469

# Rationale for this change

While working on other PRs I noticed a lot of copying and that
`ParquetMetaData` is significantly larger when the `encryption` feature
is enabled This is unnecessary and could be avoided
by using `Arc<FileDecryptionProperties>` instead of
`FileDecryptionProperties` directly.


# What changes are included in this PR?

1. Change most uses of `FileDecryptionProperties` into
`Arc<FileDecryptionProperties>`

If this is an acceptable API change, I will make a follow on PR giving
the same treatment to `FileEncryptionProperties`
(turn it into `Arc<FileEncryptionProperties>`)

# Are these changes tested?
Yes by CI


# Are there any user-facing changes?

Yes, some of the APIs now take `Arc<FileDecryptionProperties>` rather
than `FileDecryptionProperties`