Deprecate `parquet::format` module (#8615)
# Which issue does this PR close?

- part of  #8572.

# Rationale for this change
The `parquet::format` module is no longer needed, but because it has
been part of the public API it should go through the normal [deprecation
process](https://github.com/apache/arrow-rs?tab=readme-ov-file#deprecation-guidelines).

# What changes are included in this PR?

In addition to the deprecation notice, this modifies the documentation
for the `parquet-layout` binary which relies on the `format` module to
read page headers. This tool will be removed when `format` is removed.

# Are these changes tested?

None needed.

# Are there any user-facing changes?

Yes, public module is marked as deprecated. Not a breaking change,
however.