Check int96 min/max instead of panicking (#8603)
# Rationale for this change
Fixes #8614 
Part of #7806 

We currently panic if the size of an int96 isn't 12. Instead, we should
return an error message to avoid a panic.

# What changes are included in this PR?

Error message instead of panic.

# Are these changes tested?

We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
Test included, existing tests should still pass.

# Are there any user-facing changes?
None.