Fix Writer docs and rename `AvroBinaryFormat` to `AvroSoeFormat` (#8633)
# Which issue does this PR close?

- Closes #8631 

# Rationale for this change

What was originally an unframed `AvroBinaryFormat` became an
`AvroSoeFormat`. While the current implementation is correct, the
documentation and variable naming needs to be fixed prior to public
release. By doing this we reserve `AvroBinaryFormat` for a future third
writer format dedicated to unframed binary payload streaming while
providing end-users with a better API.

# What changes are included in this PR?

* Renamed `AvroBinaryFormat` to `AvroSoeFormat`
* Corrected documentation to reflect the latest state of the code.

# Are these changes tested?

No functionality/behavior was changed. Everything is covered by the
existing tests.

# Are there any user-facing changes?

* `AvroBinaryFormat` has been renamed to `AvroSoeFormat`
* Documentation has been updated as well.