[thrift-remodel] Add macro to reduce boilerplate necessary to implement Thrift serialization (#8634)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Part of #5853.

# Rationale for this change

Structures that are to be Thrift serialized as fields in another struct
must implement the `WriteThriftField` trait. In most instances the
implementation of that trait is trivial and adds unnecessary verbosity
to the code.

# What changes are included in this PR?

Add a new `write_thrift_field` macro that generates this boilerplate
code.

# Are these changes tested?

Covered by existing tests

# Are there any user-facing changes?
No