Undeprecate `ArrowWriter::into_serialized_writer` and add docs (#8621)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Related to https://github.com/apache/arrow-rs/issues/7835


# Rationale for this change


While testing the arrow 57 upgrade in DataFusion I found a few things
that need to be fixed
in parquet-rs.

- https://github.com/apache/datafusion/pull/17888

One was that the method `ArrowWriter::into_serialized_writer` was
deprecated, (which I know I suggested in
https://github.com/apache/arrow-rs/issues/8389 🤦 ). However, when
testing it turns out that the constructor of `SerializedFileWriter` does
a lot of work (like creating the parquet schema from the arrow schema
and messing with metadata)
https://github.com/apache/arrow-rs/blob/c4f0fc12199df696620c73d62523c8eef5743bf2/parquet/src/arrow/arrow_writer/mod.rs#L230-L263

Creating a `RowGroupWriterFactory` directly would involve a bunch of
code duplication

# What changes are included in this PR?

So let's not deprecate this method for now and instead add some
additional docs to guide people to the right lace


# Are these changes tested?
I tested manually upstream

# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.