[parquet] Adding counting method in thrift_enum macro to support  ENCODING_SLOTS (#8663)
# Which issue does this PR close?

- Closes [#8662]

# Rationale for this change

Related to https://github.com/apache/arrow-rs/pull/8607
We need to know how many encoding are support to create a decoder slot.

# What changes are included in this PR?
Update the `thrift_enum` to know the fields count of enum `Encoding`,
and the value is passed to `EncodingMask` And the `ENCODING_SLOTS`


# Are these changes tested?
1. Originally I think add a UT can prevent failure after the new
encoding are introduced, then I realized the counts are already
transferred, the UT is not required, the original tests can already
cover the code.

# Are there any user-facing changes?
No