perf: zero-copy path in `RowConverter::from_binary` (#8686)
# Which issue does this PR close?

- Closes #8685.

# What changes are included in this PR?

In the implementation of `RowConverter::from_binary`, the `BinaryArray`
is broken into parts and an attempt is made to convert the data buffer
into `Vec` at no copying cost with `Buffer::into_vec`. Only if this
fails, the data is copied out for a newly allocated `Vec`.

# Are these changes tested?

Passes existing tests using `RowConverter::from_binary`, which all
convert a non-shared buffer taking advantage of the optimization.
Another test is added to cover the copying path.

# Are there any user-facing changes?

No