check bit width to avoid panic in DeltaBitPackDecoder (#8688)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Part of #7806 
# Rationale for this change

The `DeltaBitPackDecoder` can panic if it encounters a bit width in the
encoded data that is larger than the bit width of the data type being
decoded.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>