[thrift-remodel] Use `thrift_enum` macro for `ConvertedType` (#8680)
# Which issue does this PR close?

- Part of #5853.

# Rationale for this change

While converting to the new Thrift model, the `ConvertedType` enum was
done manually due to the `NONE` variant, which used the discriminant of
`0`. This PR changes that to `-1` which allows the `thrift_enum` macro
to be used instead. This improves code maintainability.

# What changes are included in this PR?

See above.

# Are these changes tested?

Covered by existing tests

# Are there any user-facing changes?

No, this only changes the discriminant value for a unit variant enum.