Support more operations on ListView (#8645)
# Which issue does this PR close?

Part of #5375

Vortex was encountering some issues after we switched our preferred List
type to `ListView`, the first thing we noticed was that
`arrow_select::filter_array` would fail on ListView (and LargeListView,
though we don't use that).

This PR addresses some missing select kernel implementations for
ListView and LargeListView.

This also fixes an existing bug in the ArrayData validation for ListView
arrays that would trigger an out of bounds index panic.

# Are these changes tested?

- [x] filter_array
- [x] concat
- [x] take


# Are there any user-facing changes?

ListView/LargeListView can now be used with the `take`, `concat` and
`filter_array` kernels

You can now use the `PartialEq` to compare ListView arrays.

---------

Signed-off-by: Andrew Duffy <andrew@a10y.dev>