[Parquet] Account for FileDecryptor in ParquetMetaData heap size calculation (#8671)
# Which issue does this PR close?

- Closes #8472.

# Rationale for this change

Makes the metadata heap size calculation more accurate when reading
encrypted Parquet files, which helps to better manage caches of Parquet
metadata.

# What changes are included in this PR?

* Accounts for heap allocations related to the `FileDecryptor` in
`ParquetMetaData`
* Fixes the `HeapSize` implementation for `Arc<T>` so the size of `T` is
included, as well as the reference counts that are stored on the heap
* Fixes the heap size of type pointers within `ColumnDescriptor` being
included twice

## Not included
* Accounting for any heap allocations in a user-provided `KeyRetriever`

# Are these changes tested?

Yes, there's a new unit test added that computes the heap size with a
decryptor.

I also did a manual test that created a test Parquet file with 100
columns using per-column encryption keys, and loaded 10,000 copies of
the `ParquetMetaData` into a vector. `heaptrack` reported 1.136 GB
memory heap allocated in this test program. Prior to this change, the
sum of the metadata was reported as 879.2 MB, and afterwards it was
1.136 GB.

# Are there any user-facing changes?

No

This was co-authored by @etseidl

---------

Co-authored-by: Ed Seidl <etseidl@live.com>