Document limitations of the `arrow_integration_test` crate (#8738)
# Which issue does this PR close?

- Closes https://github.com/apache/arrow-rs/issues/8684

# Rationale for this change

The crate is apparently only intended for internal use. However, the
crate documentation does not mention this anywhere. Given that the crate
is severely incomplete in certain areas, we should document this
clearly.

# What changes are included in this PR?

- add a disclaimer to the crate-level docs that the crate is only
intended for integration-testing the Arrow project itself and not
recommended for outside use
- Document that `ArrowJsonBatch::from_batch` is deliberately incomplete
and only supports `Int8` arrays right now

# Are these changes tested?

doc-only change

# Are there any user-facing changes?

yes, the mentioned doc changes