add `try_new_with_length` constructor to `FixedSizeList` (#8624)
Closes https://github.com/apache/arrow-rs/issues/8623

As a first step for fixing #8623, this PR introduces a
`try_new_with_length` constructor for `FixedSizeListArray` that takes in
a `len` parameter for the specific case of non-nullable, degenerate
arrays.

To be honest, I think updating the existing constructors makes more
sense (breaking change), but this is an easier first step because we can
just make `try_new_with_length` the new `try_new` from here.

I also fixed 2 cases of an existing unit test that were wrong, as well
as added an extra test for the degenerate case.

Edit: I realized that the constructors also do not check if the list
size correctly divides the input `values.len()` in `try_new`, so added a
check for that plus more tests.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>