perf: Speed up Parquet file writing (10%, back to speed of 56) (#8786)
# Which issue does this PR close?

- Partial fix for #8783.

# Rationale for this change

An appreciable slowdown in Parquet writing was noticed. At least part of
the slowdown seems to stem from changes to `basic::LogicalType` which
caused cloning the enum to take make longer than previously.

# What changes are included in this PR?

This adds a new `logical_type_ref` call to `BasicTypeInfo` and
`ColumnDescriptor`. Unlike the existing `logical_type` which returns a
cloned `Option<LogicalType>`, the new methods return
`Option<&LogicalType>`. This new function is used in place of
`logical_type` internally.

# Are these changes tested?

Should be covered by existing tests.

# Are there any user-facing changes?
No.

A further optimization would be to change `ColumnOrder::get_sort_order`
to take an `Option<&LogicalType>`, but that is a breaking API change.