Fix bug in handling of empty Parquet page index structures (#8817)
# Which issue does this PR close?

- Closes #8815.

# Rationale for this change

When writing Parquet metadata, sometimes the column and offset indexes
contain missing values (this is usually a side effect of the
`ParquetMetaData` not allowing for `None` in the page index structures).
This can lead to errors or panics.

# What changes are included in this PR?

Adds some checking in `ThriftMetaDataWriter` to detect missing bits and
work around them.

# Are these changes tested?

Yes, new tests added.

# Are there any user-facing changes?

No