Allow reading of improperly constructed empty lists in Parquet metadata (#8827)
# Which issue does this PR close?

- Closes #8826.

# Rationale for this change

As reported in the issue, some writers will use an element type of 0 for
an empty list. This is not thrift compact protocol spec compliant, but
many readers (including this crate prior to 57.0.0) tolerate this.

# What changes are included in this PR?
Adds a special case to `read_list_begin` for a 0 length list with a type
of 0.

# Are these changes tested?

Yes

# Are there any user-facing changes?

No, internal change only