Respect page index policy option for ParquetObjectReader when it's not skip (#8857)
# Which issue does this PR close?

When i try to wrapper ParquetObjectReader and implement our internal
metadata cache, and i will pass the option to the inner
ParquetObjectReader, but it does not respect the index policy option
even it's not skip, and it always be false and will not load page index
which i want to prefetch and cache.

- Closes #[8856](https://github.com/apache/arrow-rs/issues/8856)

cc @alamb 

# Rationale for this change

Support option with page index if it's not skip.



# Are these changes tested?

Yes

# Are there any user-facing changes?

No

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>