Fix `RowNumberReader` when not all row groups are selected (#8863)
# Which issue does this PR close?

Closes https://github.com/apache/arrow-rs/issues/8864.

# Rationale for this change

https://github.com/apache/arrow-rs/pull/8715 introduced row numbers
feature last week. However, it had a bug, which luckily for us @scovich
pointed out soon after the merge. The issue is that the row numbers are
produced in ordinal-based order of the row groups, instead of
user-requested order of row groups. The former is wrong, and is being
fixed here by switching to user-requested order.

# What changes are included in this PR?

Just fixing the bug as explained above, and adding test. Also addressing
two small comments from post-merge review:
https://github.com/apache/arrow-rs/pull/8715#pullrequestreview-3474400469

# Are these changes tested?

Yes.

# Are there any user-facing changes?

No, this wasn't released yet.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>