Add `merge` and `merge_n` kernels (#8753)
# Which issue does this PR close?

- Closes #8752.

# Rationale for this change

The algorithms suggested in this PR originate from the `case` logic in
DataFusion (see
[datafusion#18152](https://github.com/apache/datafusion/pull/18152) and
[datafusion#18444](https://github.com/apache/datafusion/pull/18444)). I
think it might be useful to move them to `arrow-rs` instead of being
tucked away in a corner of the DataFusion codebase.

# What changes are included in this PR?

Adds a two-way and n-way merge algorithm that's halfway between `zip`
and `interleave`. In contrast to `zip` the truthy and falsy arrays do
not need to be prealigned. In contrast to `interleave` the relative
order of elements in each input array is retained in the final result.

# Are these changes tested?

I've already added two minimal unit tests, more should probably be
added.

# Are there any user-facing changes?

No breaking API changes

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>