[8910]Fixed doc test with feature prettyprint (#8911)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8910 .

# Rationale for this change

To fix the doc-test with feature-flag `prettyprint`

# What changes are included in this PR?

Enabled feature-flag `prettyprint` for specific doc-test that need this
feature.

# Are these changes tested?
Yes

# Are there any user-facing changes?
No