Fix 57.1.0 CHANGELOG to refer to virtual row number support (#8903)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Related to
https://github.com/apache/arrow-rs/pull/8874/files#r2550560291

# Rationale for this change

@vustef
[pointed](https://github.com/apache/arrow-rs/pull/8874/files#r2550560291)
out the changelog for 57.1.0 improperly misses the row number support

# What changes are included in this PR?

Fix the Changelog

NOTE this doesn't affect the changelog on the 57.1.0 tag -- I will
however, add a note to https://github.com/apache/arrow-rs/issues/8800 to
lead people there

# Are these changes tested?

CI

# Are there any user-facing changes?