feat: support casting  `Time32` to `Int64` (#8971)
# Which issue does this PR close?

- Closes https://github.com/apache/datafusion/issues/19036

While this issue was raised on the datafusion side, I think this is
where the underlying issue exists, i.e., that arrow cannot do the cast
directly.

# Rationale for this change

It should be possible to directly go from a time32 to an int64, but
currently it is not.

# What changes are included in this PR?

1. Add match arms for needed cases.
2. Add test that mimics cast issue described in datafusion issue.

# Are these changes tested?

Yes, adds test to mimic the case described in the datafusion issue.

# Are there any user-facing changes?

No