Implement `BinaryArrayType` for `&FixedSizeBinaryArray`s (#8993)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8992

# Rationale for this change

Makes `FixedSizeBinaryArray` in-line with other binary array types, and
easier for downstream users to treat them the same way.

# What changes are included in this PR?

Implement `BinaryArrayType` trait for `&FixedSizeBinaryArray`s.

# Are these changes tested?

Trait implementation only.

# Are there any user-facing changes?

New trait implementation for `&FixedSizeBinaryArray`.