Add boolean benchmark for byte aligned slices (#8997)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- part of https://github.com/apache/arrow-rs/issues/8806
- Part of https://github.com/apache/arrow-rs/pull/8996

# Rationale for this change

As part of https://github.com/apache/arrow-rs/pull/8996 I would like to
add special case
code for byte aligned boolean buffers, and to do so I would like to have
benchmarks that cover this

# What changes are included in this PR?

1. Add benchmark for offset of 24 bits (in addition to 1)

# Are these changes tested?
I ran it manually

# Are there any user-facing changes?

No