feat: implement append_array for FixedSizeBinaryBuilder (#8989)
- Implemented append_array for the builder
    - appends values and nulls one by one
    - tested with unit test
- To #8750

# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8750.

# Rationale for this change

Added `append_array` fn for the builder

# What changes are included in this PR?

There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.

# Are these changes tested?

We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?

# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.

---------

Signed-off-by: 蔡略 <cailue@apache.org>
Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>