Improve `RunArray` documentation (#9019)
# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.
-->

N/A

# Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

Whilst reviewing https://github.com/apache/datafusion/pull/18981 I found
it very confusing trying to follow along the logic for accounting for
slicing of RunArrays. Decided to see if I can improve the documentation
around it, especially to make it clear slicing acts logically not
physically.

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

Improve docstrings for RunArrays and RunEndBuffers.

Also add some tests for kernel operations on sliced RunArrays which
seemed to be missing.

# Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

Doc changes only. Added some (expected) failing tests to showcase #9018
too.

# Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.
-->

Doc changes only.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>